address comments and add changeset
