+++ .changeset/dirty-pumas-dress.md
@@ -1,0 +1,6 @@
+---
+"@aptos-labs/wallet-adapter-mui-design": major
+"@aptos-labs/wallet-adapter-nextjs-example": minor
+---
+
+Add wallet adapter MUI design to nextjs example
