+++ packages/wallet-adapter-mui-design/.gitignore
@@ -1,0 +1,1 @@
+dist
